GTM Listener response codes are buggy
Issues:
Fixes #1057

Problem:
It appears that in 11.6.0 the return code for doing a GET on a
non-existing GTM listener would return a 500 error instead of
the expected 404.

Additionally, on 12.1.2, it appears that this problem has gone
through some revisions and the return code changed to 400.

This is still incorrect though. The response code for something
not found should be 404.

Analysis:
This patch checks for the message returned by the API when an
item is not found and returns the correct value if it does
not exist

Tests:
* functional