Change proxying to be keyword arguments of the iWorkflow ManagementRoot
Issues:
Fixes #1082

Problem:
Prior to this patch, the usage was to get a ManagementProxy class. This
was not very intuitive though, so I've changed it to use keyword args
to the iWorkflow ManagementRoot class. Presumably the same would apply
to BIG-IQ

Analysis:
Refactors out ManagementProxy in favor of kwargs to ManagementRoot

Tests:
functional