Adds documentation for (manual) release process
Issues:
Fixes #1292

Problem:
The release process was not documented and if the automatic process
fails, there is no insight into how it worked in the first place,
and therefore how to fix it.

Analysis:
This patch includes docs that explain the manual process that can
be used if the automatic process fails or is not understood.

Tests: