Correct edge case in ltm node api for address + fqdn
Issues:
Fixes #1500

Problem:
The ltm node api has an edge case with the address value "any6" when
it is specified with the fqdn key.
The patch that originally fixed this issue was causing
the bigip_node module in ansible to break.

Analysis:
This patch handles that edge case and allows it.

Tests:
* functional